pyproject.toml
[project]
name = "aimbat"
dynamic = ["version"]
description = "AIMBAT: Automated and Interactive Measurement of Body-wave Arrival Times."
authors = [
    { name = "Xiaoting Lou", email = "xlou@u.northwestern.edu" },
    { name = "Simon M. Lloyd", email = "simon@slloyd.net" },
]
requires-python = ">=3.12,<3.15"
readme = "README.md"
license = "GPL-3.0-or-later"
classifiers = [
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]
dependencies = [
    "pysmo",
    "loguru>=0.7.3",
    "sqlmodel>=0.0.24",
    "rich>=13.9.4",
    "matplotlib>=3.10.6",
    "pyside6>=6.9.2",
    "pyqtgraph>=0.13.7",
    "cyclopts>=3.24.0",
    "pydantic-settings>=2.10.1",
    "pandas>=3.0.1",
    "pandas-stubs>=3.0.0.260204",
]

[project.urls]
Documentation = "https://aimbat.pysmo.org"

[project.scripts]
aimbat = "aimbat.app:app"

[dependency-groups]
test = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-dependency>=0.6.0",
    "pytest-mpl>=0.17.0",
    "pytest-mypy>=1.0.1",
    "pytest-qt>=4.5.0",
    "pytest-sugar>=1.1.1",
    "ruff>=0.13.0",
]
dev = [
    "black>=26.1.0",
]
docs = [
    "git-cliff>=2.12.0",
    "mkdocstrings-python>=2.0.2",
    "zensical>=0.0.23",
]

[tool.uv]
default-groups = [
    "test",
    "dev",
    "docs",
]

[tool.uv.sources]
pysmo = { git = "https://github.com/pysmo/pysmo", rev = "master" }

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.wheel]
packages = ["src/aimbat"]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
# xvfb_width = 1920
# xvfb_height = 1080
testpaths = [
  "tests",
  "src",
]
mpl-generate-summary = "html"
mpl-use-full-test-name = true

[tool.mypy]
mypy_path = "src"
namespace_packages = true
disallow_untyped_defs = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.coverage.run]
source = ["src"]
[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "plt.show()",
    "plt.savefig(.*)",
    "pass",
    ]

[tool.black]
target-version = [
    "py312",
    "py313",
    "py314",
]

[tool.ruff]
line-length = 88
extend-exclude = [
    "sample-data",
    "docs",
    "site",
]
